Renaming controllers name and model folder.
