+++ server/controllers/dalController.js
@@ -1,5 +1,5 @@
-const skillConfigRepoModel = require('../molels/skillConfigRepoModel').skill_config_repo;
-const taskStepUIModel = require('../molels/taskStepUIModel').taskStepUIModel;
+const skillConfigRepoModel = require('../models/skillConfigRepoModel').skill_config_repo;
+const taskStepUIModel = require('../models/taskStepUIModel').taskStepUIModel;
 
 module.exports.getUIConfigPath = function(templateId, callback) {
     getFilePath(templateId, 'ui_config_path', callback);
